All GUI files moved to GUI folder and refs refactored.
