+++ ClockheartFrontEnd/app/src/components/GUI_Files/BarterCompleteScreenGUI.js
