+++ ClockheartFrontEnd/app/src/components/GUI_Files/BarterPanelGUI.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react"
-import { updateCharacterInTable } from "./Services/CharacterServices"
-import { getPlayerItems, updateItemInTable } from "./Services/ItemServices"
+import { updateCharacterInTable } from "../Services/CharacterServices"
+import { getPlayerItems, updateItemInTable } from "../Services/ItemServices"
 
 const BarterPanelGUI = ({characters, setCharacters, items, setItems, setBarterPanel, setBarterComplete, setBarterFailed, setBarterCantAfford}) => {
