+++ ClockheartFrontEnd/app/src/components/GUI_Files/BookGUI.js
@@ -1,4 +1,4 @@
-import Book from "./Book"
+import Book from "../Book"
 
 const BookGUI = ({setBookLocationOpen}) => {
