+++ ClockheartFrontEnd/app/src/components/GUI_Files/BossDeathScreenGUI.js
