+++ ClockheartFrontEnd/app/src/components/GUI_Files/TalkCompleteScreenGUI.js
