+++ ClockheartFrontEnd/app/src/components/GUI_Files/TalkFailedScreenGUI.js
