+++ ClockheartFrontEnd/app/src/components/GUI_Files/TalkPanelGUI.js
@@ -1,4 +1,4 @@
-import { getPlayerItems, updateItemInTable } from "./Services/ItemServices"
+import { getPlayerItems, updateItemInTable } from "../Services/ItemServices"
 
 const TalkPanelGUI = ({characters, items, setItems, setTalkPanel, setFightPanel, setTalkComplete, setTalkFailed}) => {
     const newCharacters = [...characters]
